requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
albumentations=1.2.1=pypi_0
aom=3.4.0=h27087fc_1
autopep8=1.7.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2022.6.15=py38h578d9bd_0
cloudpickle=2.1.0=pyhd8ed1ab_0
cudatoolkit=11.1.1=ha002fc5_10
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.12.0=py38h0a891b7_0
dask-core=2022.7.0=pyhd8ed1ab_0
enum34=1.1.10=py38h32f6830_2
expat=2.4.8=h27087fc_0
ffmpeg=5.0.1=gpl_h512afef_107
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h8e229c2_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=h0708190_1
fsspec=2022.5.0=pyhd8ed1ab_0
gettext=0.19.8.1=h73d1719_1008
gmp=6.2.1=h58526e2_0
gnutls=3.7.6=hf3e180e_5
google=3.0.0=pypi_0
h5py=3.2.1=nompi_py38h9915d05_100
hdf5=1.10.6=nompi_h3c11f04_101
icu=70.1=h27087fc_0
imagecodecs-lite=2019.12.3=py38h71d37f0_5
imageio=2.6.1=py38_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.4=py38h43d8883_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.36.1=hea4e1c9_2
libdrm=2.4.112=h166bdaf_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=12.1.0=h8d9b700_16
libiconv=1.16=h516909a_0
libidn2=2.3.3=h166bdaf_0
libnsl=2.0.0=h7f98852_0
libpciaccess=0.16=h516909a_0
libpng=1.6.37=h753d276_3
libstdcxx-ng=12.1.0=ha89aaad_16
libtasn1=4.18.0=h166bdaf_1
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h7f98852_0
libuuid=2.32.1=h7f98852_1000
libuv=1.44.2=h166bdaf_0
libva=2.15.0=h166bdaf_0
libvpx=1.11.0=h9c3ff4c_3
libwebp-base=1.2.3=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxml2=2.9.14=h22db469_3
libzlib=1.2.12=h166bdaf_2
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h5c7f4ab_1
medpy=0.4.0=py_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
mysql-connector-python=8.0.30=pypi_0
ncurses=6.3=h27087fc_1
nettle=3.8=hc379101_0
networkx=2.8.5=pyhd8ed1ab_0
nibabel=4.0.1=pyhd8ed1ab_0
ninja=1.11.0=h924138e_0
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
olefile=0.46=pyh9f0ad1d_1
opencv-python-headless=4.5.5.64=py38_0
openh264=2.2.0=h6239696_1
openssl=3.0.5=h166bdaf_1
p11-kit=0.24.1=hc5aa10d_0
packaging=21.3=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
pathlib=1.0.1=py38h578d9bd_6
pbd=0.9=pypi_0
pillow=7.1.2=py38hb39fc2d_0
pip=22.1.2=pyhd8ed1ab_0
progressbar=2.5=pypi_0
protobuf=3.20.1=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pycodestyle=2.9.1=pypi_0
pydicom=2.3.0=pyh6c4a22f_0
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.8.13=ha86cf86_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pywavelets=1.3.0=py38h71d37f0_1
pyyaml=6.0=py38h0a891b7_4
qudida=0.0.4=pypi_0
readline=8.1.2=h0f457ee_0
scikit-image=0.19.3=py38h47df419_0
scikit-learn=1.1.2=pypi_0
scipy=1.7.3=py38hc147768_0
setuptools=63.2.0=py38h578d9bd_0
simpleitk=2.1.1=py38h3fd9d12_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.1=h4ff8645_0
svt-av1=1.1.0=h27087fc_1
threadpoolctl=3.1.0=pypi_0
tifffile=2019.7.26.2=py38_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pypi_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=0.8.0=py38
torchsummary=1.5.1=py38_0
torchvision=0.9.0=py38_cu111
tornado=6.2=py38h0a891b7_0
typing_extensions=4.3.0=pyha770c72_0
wheel=0.37.1=pyhd8ed1ab_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h7f98852_2
zlib=1.2.12=h166bdaf_2
zstd=1.4.9=ha95c52a_0
